refractor:resolved the review comments
